Upgrade dependencies, switch from govendor to Go Modules
The upgraded Terraform dependency means the provider works with Terraform 0.12.

The Go ecosystem seem to be shifting to using Modules instead, so it made sense
to switch to that here as well. It's also what Terraform suggests on
https://www.terraform.io/docs/extend/terraform-0.12-compatibility.html#upgrading-to-the-latest-terraform-sdk

Details about how Go Modules work can be found on this wiki page:
https://github.com/golang/go/wiki/Modules