Add Github actions for pull request and push (#5)
* Add Github actions for pull request and push

Add a Github workflow to run on pull requests and pushes to main.

* Rename portions of the Github Actions workflow

* Remove errcheck from Github Actions...for now

`make errcheck` is currently returning dozens of unchecked errors. Only
once each of these have been handled or an exclusion file is added should
this step be added into the build, so for now, remove it. See issue #6 for
more details.